force internet connection in news feed.
