+++ android/Culturage/app/src/main/java/com/culturage/oceans_eleven/culturage/newsFeed/NewsFeedActivity.java
@@ -12,6 +12,8 @@ import android.widget.ImageView;
 import android.widget.Toast;
 
 import com.culturage.oceans_eleven.culturage.R;
+import com.culturage.oceans_eleven.culturage.adapters.HeritageItemListAdapter;
+import com.culturage.oceans_eleven.culturage.network.NetworkUtils;
 
 public class NewsFeedActivity extends AppCompatActivity {
 
@@ -21,6 +23,7 @@ public class NewsFeedActivity extends AppCompatActivity {
 
         // Set the content of the activity to use the activity_main.xml layout file
         setContentView(R.layout.activity_fragment_work_around);
+        NetworkUtils.forceInternetConnection(NewsFeedActivity.this);
 
 
         ImageButton searchButton = (ImageButton) findViewById(R.id.searchButton);
@@ -55,7 +58,7 @@ public class NewsFeedActivity extends AppCompatActivity {
         ViewPager viewPager = (ViewPager) findViewById(R.id.viewpager);
 
         // Create an adapter that knows which fragment should be shown on each page
-        HeritageItemListAdapder adapter = new HeritageItemListAdapder(this, getSupportFragmentManager());
+        HeritageItemListAdapter adapter = new HeritageItemListAdapter(this, getSupportFragmentManager());
 
         // Set the adapter onto the view pager
         viewPager.setAdapter(adapter);
@@ -70,7 +73,6 @@ public class NewsFeedActivity extends AppCompatActivity {
         //      by calling onPageTitle()
         tabLayout.setupWithViewPager(viewPager);
     }
-
 
     @Override
     public void onBackPressed() {
