+++ android/Culturage/app/src/main/java/com/culturage/oceans_eleven/culturage/signup_login/LoginActivity.java
@@ -1,15 +1,11 @@
 package com.culturage.oceans_eleven.culturage.signup_login;
 
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.net.ConnectivityManager;
 import android.os.AsyncTask;
-import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
@@ -20,6 +16,7 @@ import android.widget.ProgressBar;
 import android.widget.Toast;
 
 import com.culturage.oceans_eleven.culturage.R;
+import com.culturage.oceans_eleven.culturage.network.NetworkUtils;
 import com.culturage.oceans_eleven.culturage.network.PostJSON;
 import com.culturage.oceans_eleven.culturage.newsFeed.NewsFeedActivity;
 
@@ -44,7 +41,7 @@ public class LoginActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_login);
-        forceInternetConnection();
+        NetworkUtils.forceInternetConnection(LoginActivity.this);
 
         preferences = PreferenceManager.getDefaultSharedPreferences(LoginActivity.this);
         String token = preferences.getString("token", "null");
@@ -73,29 +70,6 @@ public class LoginActivity extends AppCompatActivity {
         });
 
     }
-
-    private void forceInternetConnection() {
-        ConnectivityManager conMan = (ConnectivityManager) LoginActivity.this.getSystemService(Context.CONNECTIVITY_SERVICE);
-        if (conMan.getActiveNetworkInfo() == null || !conMan.getActiveNetworkInfo().isConnected()) {
-            AlertDialog.Builder builder;
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                builder = new AlertDialog.Builder(LoginActivity.this, android.R.style.Theme_Material_Dialog_Alert);
-            } else {
-                builder = new AlertDialog.Builder(LoginActivity.this);
-            }
-            builder.setTitle("No Internet Connection")
-                    .setMessage("Please connect to the internet")
-                    .setPositiveButton(android.R.string.yes, new DialogInterface.OnClickListener() {
-                        public void onClick(DialogInterface dialog, int which) {
-                            finish();
-                        }
-                    })
-                    .setIcon(android.R.drawable.ic_dialog_alert)
-                    .show();
-        }
-    }
-
-
     private void attemptLogin() {
 
         // Reset errors.
